+++ NEWS.md
@@ -1,6 +1,7 @@
 # huito 0.1.1
 
 - Update vignettes.
+- Update webpage
 - Match argument for `label_print(mode)` to "preview" and "sample".
 
 # huito 0.1.0