+++ NEWS.md
@@ -3,6 +3,7 @@
 - Update vignettes.
 - Update webpage
 - Match argument for `label_print(mode)` to "preview" and "sample".
+- Fix the number of label to print: `label_print(nlabels)`
 
 # huito 0.1.0