+++ NEWS.md
@@ -1,9 +1,10 @@
 # huito 0.1.1
 
+- Update pkgdown
 - Update vignettes.
-- Update webpage
+- Include example from GerminaR sticker 
 - Match argument for `label_print(mode)` to "preview" and "sample".
-- Fix the number of label to print: `label_print(nlabels)`
+- Set the number of label to print: `label_print(nlabels)`
 
 # huito 0.1.0